requirements.txt
tensorflow==2.8.0
tensorflow-hub==0.12.0
pillow==9.0.1
streamlit==1.8.1
matplotlib==3.5.1
numpy==1.21.5
requests==2.27.1
